Updated The Script along with some bug fixes

+++ New folder/Laptop Database System/Laptop Database System/Controller.cs
@@ -103,7 +103,7 @@ namespace Laptop_Database_System
             return dbMan.ExecuteNonQuery(query);
         }
 
-        public int addStore(string name,int number,string address)
+        public int addStore(string name,long number,string address)
         {
             string query = "INSERT INTO Store Values('"+name+"',"+number+",'"+address+"')";
             return dbMan.ExecuteNonQuery(query);
